# Make selectivity vector optional for DecodedVector