--- a/velox/functions/prestosql/benchmarks/CMakeLists.txt
+++ b/velox/functions/prestosql/benchmarks/CMakeLists.txt
@@ -156,3 +156,8 @@ add_executable(velox_functions_prestosql_benchmarks_map_zip_with
                MapZipWithBenchmark.cpp)
 target_link_libraries(velox_functions_prestosql_benchmarks_map_zip_with
                       ${BENCHMARK_DEPENDENCIES})
+
+add_executable(velox_functions_prestosql_benchmarks_cardinality
+               CardinalityBenchmark.cpp)
+target_link_libraries(velox_functions_prestosql_benchmarks_cardinality
+                      ${BENCHMARK_DEPENDENCIES})