Summary:

When we want to decode a vector for all its rows, we must create a 
selectivity vector of size (vector. size()). The construction of this vector is
 much more expensive than decoding. In practice, we can avoid that. 
For functions like cardinality, when we never access the elements, the 
decoding cost of the elements becomes expensive because of that.

This diff makes rows optional in the DecodedVector interface.
The vector is decoded for all rows [0, size()) when not passed.

##cardinality
before:
```
============================================================================
[...]l/benchmarks/CardinalityBenchmark.cpp     relative  time/iter   iters/s
============================================================================
simple                                                     12.28ms     81.43
vector                                                      5.55ms    180.03
fast                                                        5.02ms    199.24
simpleDictionary                                           38.97ms     25.66
vectorDictionary                                           42.46ms     23.55
simpleConditional                                          29.85ms     33.50
vectorConditional                                          18.28ms     54.72
```
after:
```
============================================================================
[...]l/benchmarks/CardinalityBenchmark.cpp     relative  time/iter   iters/s
============================================================================
simple                                                      5.37ms    186.07
vector                                                      5.31ms    188.40
fast                                                        4.79ms    208.89
simpleDictionary                                           36.72ms     27.23
vectorDictionary                                           36.42ms     27.46
simpleConditional                                          19.01ms     52.60
vectorConditional                                          18.05ms     55.39                                   1.25ns   803.10M

```